Isolate root model, implement project/user counts
Now the root page model is no longer special; instead, anything that
needs to be shared throughout the application (like currently logged
in user) is in a React context. This means we can simplify our pages
and increase their isolation.

We also implement the API to count projects and non-test projects on
the server, and have the root page call it.