Switch from npm to yarn
NPM's package-lock.json is not a true lockfile, as `npm install` can
produce different results on different machines even though
package-lock.json is committed to the repo(!). Yarn has a true lockfile
and produces deterministic, consistent results across machines.

Note that this does not *entirely* remove the use of npm in this
project, as the UseAngularCliServer class is hardcoded to use "npm run"
to run scripts from package.json. But "npm run" will not cause lockfile
conflict issues, so this is not a true problem.