Invert meaning of bool parameter for projects
Previously the bool parameter was named "wantPrivate", and was true for
requesting a private project but false for requesting a public one. It
makes more sense for the parameter to match the database's is_public
column, so the parameter is now "isPublic" and means true for a public
project, false for a private one. Far less confusing in the long run.