Implement project deletion (really archiving)
Since we're matching the Redmine database schema here, we reuse their
project status values (1 for active, 5 for closed, 9 for archived) even
though we don't need to distinguish between active and closed so it
would make more sense (for us) to have a Boolean is_archived field.
But this way no schema changes are needed.