Update API for Redmine 4 data model
Of the changes to Redmine's data model between versions 0.8 and 4.x,
there are just two that affect us:
* Users can now have multiple email addresses, so mail_address is a
  separate table and the users table doesn't have a Mail field
* Users can now have multiple roles in a project, so the members table
  no longer contains a role_id field; instead, there's a member_roles
  table that tracks the many-to-many relation between members and roles.

Next up: redesigning the API so that the frontend can handle multiple
emails for a single user, and also trimming a lot of the Redmine fields
that we don't care about from the API, so that the API only has to care
about things like username, real name, password, and email. The rest of
the Redmine fields we'll just leave alone, or create records with them
set to their default values.