WIP: Rewrite internal model API
Now the model, and the in-memory storage model (used for unit tests),
both expect functions that use the new API design. Controller hasn't
been updated yet, nor has the actual guts of the model code, so this
won't compile. But the shape of the internal API is much better now,
which is reflected in how much simpler MemoryStorage is (and half its
code has been moved to MemoryModel).