WIP: Can now remove user from all roles in project
This finishes up a couple of TODOs. We also now populate project
membership in the FromSql function for ProjectDetails, and populate the
list of emails (with isDefault one(s) first in the list) in the FromSql
function for UserDetails, which was another TODO.

Also in this commit: whitespace improvements in query expressions.