Add a bit of test data and fix SQL in test data
One SQL statement in the test data had no actual data attached, so I
commented that one out as it was a syntax error to have an INSERT INTO
statement with no values. Also added one member role so that the "test"
user will be a member of one project with two different roles (manager
and programmer), to allow testing more scenarios.