WIP: Back to SQL model; slightly better project page
Now the project page will list membership, but only for single projects.
Still to do: getting membership for a whole list of projects (project
code in (this list)) from SQL, so that the page where "which projects am
I a member of?" will look good.

Note that there seems to be some weirdness with List.isEmpty here, that
I haven't figured out yet.