Tweak test data a little bit
Now the "LD API UTF8 Eñcoding" project has at least one member so we can
check that its membership records get pulled correctly, and one user has
two email addresses, with the default one coming second in the MySQL
database (at least when ordered by the table's Id column) so that we can
check that the sorting works correctly as well.

SampleData.fs has also been updated with the same changes, so that if we
use MemoryModel we'll get the same result.