WIP: Bump version of SQLProvider to fix SQL bug
The cause of the problem I was having earlier with the MySQL model was a
bug in SQLProvider where the order of the "where" and "join" clauses
mattered: if a "where" clause was before the "join" clause it would end
up not returning data for that table (but would still return data for
the other table in the join). With the most recent SQLProvider version,
that bug is fixed and I can now put the "where" clauses earlier in the
query, closer to the table that they're filtering. This has no effect on
the actual SQL produced, but it reads better in the F# code.