Temporarily stop copying MySQL DLLs
Copying the MySQL DLLs into place seems to be messing with running
instances of VS Code, leading to a lot of error messages about the
running F# code completion server being unable to access certain DLLs.
Revert this commit once we figure the issue out, as this is not
something we want to do long-term.

This commit also includes an extra line to print SQL commands out to the
console as they get run, for help in debugging.