Better indentation in blockinfile module
YAML operators, including the | block-indentation operator, can take an
optional number (from 1 to 9) after the operator that indicates how many
spaces of indentation to "swallow" from the block in addition to the
indentation of the parent line. Any remaining spaces will be included in
the block. So by doing |2 and indenting six spaces from the parent, we
end up with four spaces of indentation in the Apache config file,
matching the indentation of the part of that config file that this block
belongs to. Result: a better-looking config file.