Switch database name to match production name
Since the SqlProvider code has the database name as part of the type
name, we should use the same name that will be used on the QA and live
servers, rather than a different database name.

TODO: Find out how to tell SqlProvider to *not* include the database
name in the type system, because we need to be able to use the same code
with both the public and private databases, which have different names.