Simpler way to handle arguments in build script
The FAKE build system now has a simpler way to get extra arguments
passed to your build target: each target runner function gets a
parameter that includes a context, and the context includes (among other
things) the list of extra arguments passed to this target on the command
line. That's nicer than a global variable.