Disable creation of site.map file in server
The site.map file is used by the Saturn framework tooling to help debug
routing, but we don't want it in production. In theory we could wrap
this line with #if DEBUG (nothing) #else disable_diagnostics #endif, but
in practice we never use that file anyway so we'll just turn it off.