Use different DB names for public and private
The problem with the Owner parameter in SqlProvider is that it sets the
database name, overriding the one from the connection name, meaning that
MySQL, at least, will fetch from the public database even if the
connection string was set to the private database. (Other DB providers
might have different behavior; in Postgres, for example, a schema is a
separate concept from a database, and Owner sets the schema). So we need
to stop using the Owner parameter in order to get public and private DB
connections to work the way we want them to.