Switch to using manually-written SQL statements
Until https://github.com/fsprojects/SQLProvider/issues/501 is fixed, we
won't be able to use the SQLProvider library to do our queries, because
it has an unfortunate interaction with MySQL's lack of support for true
SQL "schemas" (which should really be called namespaces). Details in the
linked GitHub issue so I won't repeat them here.