WIP on changing how we return membership in JSON
Now that we're only allowing one role per project, the way the DTO
handles membership is greatly simplified. We'll take what we get out of
MySQL and pass it along directly as a JSON "object" (dictionary) of
multiple { username: role } keys, where both username and role are
strings. This should make the client code much easier to write as well.