No longer require logging in to list projects
We're going to move the server to a mode where you use an API token to
access everything. (It will be a shared-secret token, where using the
token gets you access to the entire API and the token is kept on the
build server). So there's no need to verify login details in the
project-membership API any more.