Slightly less complicated API JSON return type
Due to https://github.com/thoth-org/Thoth.Json/issues/49 we're having
trouble with the JsonSuccess / JsonFailure type. So we'll be slightly
less complicated in the type we return: there will always be data and
message fields, but on success the message field will be empty and on
failure the data field will be null, or undefined, or 0, or some kind
of "empty" value (depending on the type). This lets us avoid the type
issue that was causing failures earlier.