Now front page can load users, projects, and roles
The data-table component I wrote is quite versatile. It *cannot* switch
its table definition on the fly, though, so I need to use ngIf or
ngSwitch to switch in or out the appropriate component for each query.
But that's not going to be a problem in the actual app itself.