Add routing and user & project components
Single-user component is just a proof-of-concept demo here, and it uses
the wrong approach (see comment in single-user.component.html). But the
routing part works, and that's what I'm focusing on for now.

Single-project component is also just a proof-of-concept demo, but it
uses the correct approach, as opposed to single-user component. I've
left single-user component as-is in this comment for demo purposes.