Show one way to fix the single-user page
We can fix the single-user page emitting four HTTP API requests by using
a Subject so that it will send the same value to any subscription (since
each instance of AsyncPipe creates a new subscription under the hood).
Actually we want to use a BehaviorSubject so that if we end up
subscribing after the HTTP request returns, we still get the value.