Add auth0 login service
Started with the default Angular SPA service that auth0 suggests, and
added a couple of BehaviorSubjects to track the *current* login and user
profile status at all times, so that passive components like the profile
component won't have to trigger a new API call. I also replaced the
hardcoded calls to `window.location` with a LocationService so that we
can replace it with a Jasmine spy in testing.