Register service classes rather than functions
Instead of creating a type for every different function we call, we'll
just register a single class and move the functions to be class methods.
Makes the component registration *much* simpler and gets rid of the need
to distinguish the ListUsers type from ListProjects (and so on).

This will also allow us to simplify the public/private distinction, by
creating controllers that will work on either one and just pull the
appropriate type of model out of the service collection.