List APIs now take query params limit and offset
Now instead of specifying /limit/%d or /offset/%d in the URL, limit and
offset will be query params. This makes the URL more REST-like because
it means "the complete list of users" no matter what, but some queries
might be interested in getting just *part* of the list, which is what
query params are actually for.