Demo of dealing with API project membership data
API returns membership data as an object with keys being usernames and
values being member roles. This is a good design for the API, but if the
Typescript code wants the membership data in a different shape, then the
project model is the right place to convert it. This commit shows one
good way to convert the data returned by the API into a different shape
that we might want to use.