Generalize the user typeahead
The only problem with passing callback functions into a component as
inputs is that you have to write them as "myFunc.bind(this)" in the
template, which isn't quite the most elegant way to do it.

The official Angular way, of course, would be for the typeahead
component to just have an event emitted whenever its contents change,
and the parent is the one that hooks it up to the API call and
subscribes to the results. But I wanted to show a way to let the
component say "Hey, you tell me how to fetch data, and I'll just give
you back a stream of new data". I.e., let the component take a function
as parameter in classic functional-programming style. And it works.