Use arrays in routerLink values
As suggested in the Angular documentation, this can be a better way to
specify the link since it lets Angular build it up for us. Here it's not
a big deal, but with more complicated links it might be far easier with
arrays than building the link ourselves inside the template.