Server, not model, will decide API authorization
Either the API server itself, or else the Apache reverse proxy server,
will allow only connections from certain IP addresses to go through. So
the model is not the right place to be making "is this request allowed?"
decisions. By the time the request reaches the MySQLModel implementation
we know it's authorized, so just accept it.