Membership model and test routes
Note that the Membership model needs to be in the same file as the
Project and User models to avoid circular import loops. The objection.js
library has other methods to avoid circular import loops, but since it
uses CJS require() calls to load models when you give it pathnames, this
doesn't work with Snowpack which needs ESM import() calls. And since
import() calls are async and objection.js expects sync, it doesn't line
up. But putting all the models in one file, while a little verbose,
solves the circular import problem.