Make jump-to-definition work in VS Code
The "allowJs" config option in tsconfig.json makes the Typescript
compiler (and more importantly, the language server) process Javascript
files. This means that the "paths" section will be processed, so the
language server will know what a "$utils" import means. So now a
ctrl-click on a function call will be able to jump to its definition.