Admins can see email addresses when searching users
Non-admins only get to see the user data (name, username, and so on) but
not email addresses. Note that project managers will be able to see
email addresses of users in their project, and we don't require
confirmation before adding someone to the project, so this isn't a
complete privacy guard, because a project manager will still be able to
find out someone's email address by adding them to the project. But it's
good enough to cover the obvious holes, and we can always restrict
project managers from seeing members' email addresses later if that
becomes necessary.