Allow specifying roles with either name or ID
If the role is a number, or a string that looks like a positive integer
(no minus sign or decimal point), then it will be treated as a role ID.
Roles that do not look like numbers will be treated as role names
(manager, contributor, and so on). This gives client code more
flexibility in calling the API: assuming role IDs are well-known and
don't change, the client code won't need to know what name the API has
assigned to the role.