Fix date handling in insert, update endpoints
MySQL doesn't like the trailing "Z" suffix that most JSON producers are
going to send dates with (standard ISO format for UTC), so we need to
parse the incoming date fields if they're strings. Fortunately there's a
good library out there that is tree-shaking friendly so it will only add
the one function we use to the final build.