Move Role model into models module
The rest of the models are in one file, and it doesn't make sense for
the Role model to have its own separate file. It's an artifact of when I
was keeping each model in its own file, but the relationMappings code
ended up needing every other model in one file. So let's be consistent.