Add MYSQL_DATABASE_PRIVATE config key
Now the MySQL database name for the private instance will default to
adding "pvt" to the end of the public database name, but can be
explicitly set by setting MYSQL_DATABASE_PRIVATE or MYSQL_DATABASE_PVT
keys in the environment. If MYSQL_DATABASE is not set, the private
database instance will default to "testldapi" just like the public
instance, to allow for smoother development.