Return JSON 404 instead of HTML for unknown routes
Unknown routes (API endpoints) will now return a JSON 404 error instead
of the default Svelte-Kit 404 page, which is HTML and causes problems
when downstream clients try to parse it as JSON.

This works because "spread" routes (which capture the entire remaining
URL) have a lower match priority than non-spread routes. So anything
that can match a more specific API path will do so, and the "spread"
route is fallback that matches only when nothing else can match.