API 404 should be JSON, not HTML (#13)
Unknown routes (API endpoints) will now return a JSON 404 error instead
of the default Svelte-Kit 404 page, which is HTML and causes problems
when downstream clients try to parse it as JSON.

This works because "spread" routes (which capture the entire remaining
URL) have a lower match priority than non-spread routes. So anything
that can match a more specific API path will do so, and the "spread"
route is fallback that matches only when nothing else can match.

Also upgraded Svelte-Kit and Svelte dependencies, and changed a few
lines in package.json since Svelte and Svelte-Kit now need to be in
the dependencies part, not in devDependencies.