Fix wrong start row in Mercurial log viewer
The SVG is placed in the table body, but the table body's first row
includes the header rows in its offsetTop value. So we need to calculate
the height of the header row and subtract that from the calculated Y
positions in the SVG. Currently we don't pass in the header row at all,
so at the moment we just guess that the header and the first row are the
same size, which will only be true if the first row doesn't wrap its
text. A proper bugfix will involve passing in the header row or its
height so that we can calculate the actually-correct values.