Fix parent links in Mercurial log viewer
If a child commit had already registered its parent's column, but later
on we relocated the parent's column for a smoother log, this could cause
the child's link to the parent to be drawn in the wrong place. By not
passing parent-column data into the SVG graph component, we ensure that
by the time the parent X coordinate is calculated, the parent's location
is set and will not be moved again. This fixes the parent links that had
been exhibiting this particular bug.