Delete two TODO comments that are already done
These two TODO comments mention the need to check the JWT, but that's
already done in the calling function. There's no need to double-check it
in the implementation method. Note that createUser() is different: here
the JWT can't be checked in the calling function because it's not until
we get the DB results back that we know if we're creating or updating a
user; anyone may create a user without having a JWT, so the calling
function can't do the checks and has to defer those checks to createUser.