Slim down Docker image size by reusing layers
Until now we've had the Docker image's version number in package.json,
but that means that every time we bump the version number, it's a new
package.json and therefore the Docker build process has to rerun the
`pnpm i` step even though there were no packaging changes. By moving the
ldapi version number out of package.json into a separate file, I can
ensure that the large layer containing node_modules is only re-created
in the Docker image when it actually changes.

This will still end up creating that new layer one more time due to the
package.json change required to make this work, but that will be the
last time the node_modules layer has to be created for a version bump.