Bump version number
Removing API endpoints would be a major version number if after 1.0, but
we're in pre-1.0 versions so it's a minor version number bump.

Note that I'm *not* changing the version number in package.json, so that
the Docker image won't rebuild the node_modules layer. From now on the
package.json version number will stay constant.