Add API endpoint for "is user a project member?" (#22)
Basically, it's the GET /api/v2/projects/{projectCode}/user/{username}
endpoint, but as a HEAD request, so it only returns 200 if the user is a
member, or 404 if the membership/user/project is not found. Same
security requirements as the GET request: only that user, a project
manager, or a site admin is allowed to check this.