Update to current Svelte-Kit and Playwright
This replaces Mocha tests with Playwright API-based tests, so we have a
more consistent testing story. It also updates to the current Svelte-Kit
version since there have been a lot of changes since the last update of
the Language Depot API. One in particular, the change from lowercase to
uppercase names of the HTTP method handler functions, required a rewrite
of most of the code.

Tests were all converted, and all still pass.